+++ L_DelayLight.lua
@@ -10,9 +10,9 @@ module("L_DelayLight", package.seeall)
 local debugMode = false
 
 local _PLUGIN_NAME = "DelayLight"
-local _PLUGIN_VERSION = "1.4"
+local _PLUGIN_VERSION = "1.5"
 local _PLUGIN_URL = "https://www.toggledbits.com/delaylight"
-local _CONFIGVERSION = 00107
+local _CONFIGVERSION = 00109
 
 local MYSID = "urn:toggledbits-com:serviceId:DelayLight"
 local MYTYPE = "urn:schemas-toggledbits-com:device:DelayLight:1"
@@ -106,7 +106,7 @@ local function checkVersion(dev)
     if isOpenLuup then 
         local s = luup.variable_get( "openLuup", "Version", 2 ) -- hardcoded device?
         local y,m,d = string.match( s or "0.0.0", "^v(%d+)%.(%d+)%.(%d+)" )
-        local y = tonumber(y) * 10000 + tonumber(m)*100 + tonumber(d)
+        y = tonumber(y) * 10000 + tonumber(m)*100 + tonumber(d)
         D("checkVersion() checking openLuup version=%1 (numeric %2)", s, y)
         if y < 180400 or y >= 180611 then return true end -- See Github issue #5
         L({level=1,msg="openLuup version %1 is not supported. Please upgrade openLuup. See Github issue #5."}, y);
@@ -308,6 +308,17 @@ local function isSensorType( dev, obj, vtDev )
     return luup.device_supports_service(SENSOR_SID, dev)
 end
 
+-- Return true if device is dimmable
+local function isDimmerType( dev, obj, vtDev )
+    assert(type(dev) == "number")
+    obj = obj or luup.devices[dev]
+    if obj == nil then return false end
+    -- Can be recognized by category, implying the device is an actual switch or dimmer
+    if obj.category_num == 2 then return true end
+    -- Devices that implement dimmer behavior may be considered dimmers
+    return luup.device_supports_service(DIMMER_SID, dev)
+end
+
 -- Return true if we can treat the device as a switch
 local function isSwitchType( dev, obj, vtDev )
     assert(type(dev) == "number")
@@ -399,6 +410,13 @@ local function watchMap( m, tdev )
                 ix.variable = "Tripped"
                 ix.valueOn = "1"
                 ix.watched = true
+            elseif isDimmerType( nn, nil, tdev ) then -- dimmer/light
+                D("watchTriggers(): watching %1 (%2) as dimmer", nn, luup.devices[nn].description)
+                watchVariable( DIMMER_SID, "LoadLevelStatus", nn, tdev)
+                ix.service = DIMMER_SID
+                ix.variable = "LoadLevelStatus"
+                ix.valueOn = { { comparison=">", value=0 } }
+                ix.watched = true
             elseif isSwitchType( nn, nil, tdev ) then -- light or switch
                 D("watchTriggers(): watching %1 (%2) as switch", nn, luup.devices[nn].description)
                 watchVariable( SWITCH_SID, "Status", nn, tdev )
@@ -411,7 +429,7 @@ local function watchMap( m, tdev )
                 watchVariable( TIMERSID, "Timing", nn, tdev )
                 ix.service = TIMERSID
                 ix.variable = "Timing"
-                ix.valueOn = "1"
+                ix.valueOn = { { comparison=">", value=0 } }
                 ix.watched = true
             else
                 L({level=2,msg="Device %3 %1 (%2) doesn't seem to be a sensor or controllable load. Ignoring."},
@@ -563,16 +581,37 @@ local function isDeviceOn( devnum, dinfo, newVal, tdev )
 
     -- dinfo, which contains the trigger map data for this device, as enough information that we can use
     -- it exclusively to see what our device state is.
-    local testValues = dinfo.valueOn or "1"
+    local testValues = dinfo.valueOn or { { comparison="!=", value=0 } }
     if type(testValues) ~= "table" then testValues = { testValues } end
     -- Get inversion state
     local inv = dinfo.invert
-    -- For inequality, invert the inverted state :-)
-    if dinfo.comparison == '<>' or dinfo.comparison == "~=" or dinfo.comparison == "!=" then inv = not inv end
-    D("isDeviceOn() testing %1 val %2 against %3", devnum, newVal, testValues)
+    local nVal = tonumber( newVal ) or 0
+    D("isDeviceOn() testing %1 val %2 against %3 (invert=%4)", devnum, newVal, testValues, inv)
     for _,tv in ipairs( testValues ) do
-        if iif( inv, newVal ~= tostring( tv ), newVal == tostring( tv ) ) then
-            return true
+        local op, val
+        if type( tv ) == "table" then
+            val = tv.value or 1
+            op = tv.comparison or "="
+        else
+            val = tv
+            op = "="
+        end
+        D("isDeviceOn() dev %1 checking %2 %3 %4", devnum, newVal, op, val)
+        if op == "" or op == "=" then
+            if iif( inv, newVal ~= tostring( val ), newVal == tostring( val ) ) then return true end
+        elseif op == "!=" then
+            if iif( inv, newVal == tostring( val ), newVal ~= tostring( val ) ) then return true end
+        elseif op == ">" then
+            if iif( inv, nVal <= tonumber( val ), nVal > tonumber( val ) ) then return true end
+        elseif op == "<" then
+            if iif( inv, nVal >= tonumber( val ), nVal < tonumber( val ) ) then return true end
+        elseif op == "in" then
+            for _,v in pairs( split( val, "," ) or {} ) do
+                if v == newVal then return true end
+            end
+        else
+            L({level=2,msg="%1 (%2) device 'on' condition %3 invalid, skipping (bug, please report)."},
+                tdev, luup.devices[tdev].description, tv)
         end
     end
     return false
@@ -800,16 +839,25 @@ local function timer_runOnce( tdev )
             luup.variable_set( TIMERSID, "HouseModes", "", tdev )
             luup.variable_set( TIMERSID, "InhibitDevices", "", tdev )
             luup.variable_set( TIMERSID, "ActivePeriods", "", tdev )
+            luup.variable_set( TIMERSID, "ManualOnScene", "1", tdev )
+            luup.variable_set( TIMERSID, "ResettableOnDelay", "1", tdev )
         end
         luup.variable_set( TIMERSID, "Version", _CONFIGVERSION, tdev )
         return
     end
 
     -- Consider per-version changes.
+    L("%2 (%1) applying changes to config up to rev %3", tdev, luup.devices[tdev].description, _CONFIGVERSION)
     if s < 00106 then
-        L("Upgrading timer %1 (%2) to 00106", tdev, luup.devices[tdev].description)
         luup.variable_set( TIMERSID, "InhibitDevices", "", tdev )
         luup.variable_set( TIMERSID, "ActivePeriods", "", tdev )
+    end
+    -- Nothing for timers in 00107, master device only.
+    if s < 00108 then
+        luup.variable_set( TIMERSID, "ManualOnScene", "1", tdev )
+    end
+    if s < 00109 then
+        luup.variable_set( TIMERSID, "ResettableOnDelay", "1", tdev )
     end
 
     -- Update version last.
@@ -838,8 +886,8 @@ local function plugin_runOnce( pdev )
     end
 
     -- Consider per-version changes.
+    L("Applying config upgrades to plugin to version %1", _CONFIGVERSION)
     if s < 00103 then
-        L("Upgrading plugin %1 to 00103...", pdev)
         -- Conversion to 00103. Find all DLs incl this one and create a child
         -- of this one for it. Link it via the OldDevice state variable, which
         -- we'll detect separately.
@@ -866,16 +914,12 @@ local function plugin_runOnce( pdev )
         L("RELOADING LUUP!")
         luup.reload()
     end
-
     if s < 00105 then
-        L("Upgrading plugin %1 configuration to 00105...", pdev)
         luup.variable_set(MYSID, "NumChildren", 0, pdev)
         luup.variable_set(MYSID, "NumRunning", 0, pdev)
         luup.variable_set(MYSID, "Message", "", pdev)
     end
-    
     if s < 00107 then
-        L("Upgrading plugin %1 configuration to 00107...", pdev)
         luup.variable_set(MYSID, "DebugMode", 0, pdev)
     end
     
@@ -889,14 +933,21 @@ end
 function addTimer( pdev )
     D("addTimer(%1)", pdev)
     local ptr = luup.chdev.start( pdev )
+    local highd = 0
+    luup.variable_set( MYSID, "Message", "Adding timer. Please hard-refresh your browser.", pdev )
     for _,v in pairs(luup.devices) do
         if v.device_type == TIMERTYPE and v.device_num_parent == pdev then
+            D("addTimer() appending existing device %1 (%2)", v.id, v.description)
+            local dd = tonumber( string.match( v.id, "t(%d+)" ) )
+            if dd == nil then highd = highd + 1 elseif dd > highd then highd = dd end
             luup.chdev.append( pdev, ptr, v.id, v.description, "",
                 "D_DelayLightTimer.xml", "", "", false )
         end
     end
-    luup.chdev.append( pdev, ptr, "t"..os.time(), "DelayLight Timer", "",
-        "D_DelayLightTimer.xml", "", TIMERSID .. ",Version=0", false )
+    highd = highd + 1
+    D("addTimer() creating child d%1t%2", pdev, highd)
+    luup.chdev.append( pdev, ptr, string.format("d%dt%d", pdev, highd),
+        "DelayLight Timer " .. highd, "", "D_DelayLightTimer.xml", "", "", false )
     luup.chdev.sync( pdev, ptr )
     -- Should cause reload immediately.
 end
@@ -945,6 +996,7 @@ local function trigger( state, tdev )
     local onDelay = 0
     if status == STATE_IDLE then
         -- Trigger from idle state
+        local timing = 1
         if state == STATE_AUTO then
             if not isActiveHouseMode( tdev ) then
                 D("trigger() not in an active house mode, not triggering")
@@ -958,6 +1010,7 @@ local function trigger( state, tdev )
                 luup.variable_set( TIMERSID, "OnTime", 0, tdev, TIMERSID )
             else
                 luup.variable_set( TIMERSID, "OnTime", os.time() + onDelay, tdev )
+                timing = 2
                 D("trigger() configuring on delay %1 seconds", onDelay)
             end
         else
@@ -967,14 +1020,20 @@ local function trigger( state, tdev )
             luup.variable_set( TIMERSID, "OnTime", 0, tdev )
         end
         luup.variable_set( TIMERSID, "Status", state, tdev )
-        luup.variable_set( TIMERSID, "Timing", 1, tdev )
+        luup.variable_set( TIMERSID, "Timing", timing, tdev )
         luup.variable_set( TIMERSID, "OffTime", os.time() + onDelay + offDelay, tdev )
         scheduleDelay( scaleNextTick( onDelay + offDelay ), tdev )
 
-        -- Finally, if there's no onDelay, turn loads on. Do this last, so status
-        -- is properly set so when watches start coming for devices, the watch knows
-        -- it's a reaction to auto mode, not a manual start.
-        if onDelay == 0 then
+        -- Finally, if there's no onDelay, turn loads on. Do this last, 
+        -- so Status is properly set so when watches start coming for devices,
+        -- the watch knows it's a reaction to auto mode, not a manual start.
+        -- Prior to 1.5, a manual trigger would force all "on" devices to their
+        -- configured conditions (so switching one light on turns on all). As
+        -- of 1.5, this remains the default, but can be changed to leave devices
+        -- alone (for manual only) by setting ManualOnScene=0.
+        if onDelay == 0 and
+            ( state == STATE_AUTO or getVarNumeric( "ManualOnScene", 1, tdev, TIMERSID) ~= 0 )
+            then
             doLoadsOn( tdev )
         end
     else
@@ -1254,6 +1313,8 @@ local function timerTick(tdev)
             else
                 L("Timer %1 (%2) end of on delay, turning loads on.", tdev, luup.devices[tdev].description)
                 luup.variable_set( TIMERSID, "OnTime", 0, tdev )
+                luup.variable_set( TIMERSID, "Timing", 1, tdev )
+                -- N.B.: Since OnDelay/OnTime only applies to AUTO, no concern about forcing load states here.
                 doLoadsOn( tdev )
                 local delay = offTime - now
                 setMessage( "Delay Off " .. formatTime(delay), tdev )
@@ -1424,13 +1485,19 @@ local function timerWatch( dev, sid, var, oldVal, newVal, tdev, pdev )
             elseif status ~= STATE_IDLE then
                 -- Tripped trigger device is resetting in active state
                 local holdOn = getVarNumeric( "HoldOn", 0, tdev, TIMERSID )
-                local offTime = getVarNumeric( "OffTime", 0, tdev, TIMERSID )
-                D("timerWatch() trigger reset in %4, HoldOn=%1, offTime=%2, passed=%3", holdOn, offTime, offTime <= os.time(), status)
+                D("timerWatch() trigger reset in %1, HoldOn=%2", status, holdOn)
                 local left, ldev = isAnyTriggerOn( true, false, tdev )
-                if holdOn ~= 0 then
+                local onTime = getVarNumeric( "OnTime", 0, tdev, TIMERSID )
+                if not left and onTime > 0 and getVarNumeric("ResettableOnDelay", 1, tdev, TIMERSID) ~= 0 then
+                    -- All triggers reset during on-delay. Reset timer.
+                    L("Timer %1 (%2) all triggers reset during on-delay; resetting.",
+                        tdev, luup.devices[tdev].description)
+                    resetTimer( tdev )
+                elseif holdOn ~= 0 then
                     if not left then
-                        -- Now no more trigger devices on
-                        L("Timer %1 (%2) end of hold (%3)", tdev, myname, holdOn)
+                        -- Now no more trigger devices on in hold-over (mode 1 or 2)
+                        local offTime = getVarNumeric( "OffTime", 0, tdev, TIMERSID )
+                        L("Timer %1 (%2) end of hold (%3); offTime=%4", tdev, myname, holdOn, offTime)
                         if holdOn == 2 then
                             -- When HoldOn = 2, extend off time beyond untrip.
                             trigger( status, tdev )